Moved `guards` module to the `common` folder.
