+++ chalice_restful/common/guards.py
